[TIMOB-2674] Whoops, shutil.rmtree requires a sanity check.

+++ support/iphone/transport.py
@@ -118,7 +118,9 @@ def main(args):
 		
 		#... But we still have to nuke the stuff that gets built that we don't want
 		# to bundle.
-		shutil.rmtree(os.path.join(build_dir,'build'))
+		ios_build = os.path.join(build_dir,'build')
+		if os.isdir(ios_build):
+			shutil.rmtree(os.path.join(build_dir,'build'))
 		
 		# Install applogo/splash/etc.
 		info("Copying icons and splash...")
